+++ src/main/java/ru/fisher/VehiclePark/dto/TripImportData.java
@@ -3,6 +3,8 @@ package ru.fisher.VehiclePark.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class TripImportData {
@@ -15,6 +17,6 @@ public class TripImportData {
     private Double startLongitude;
     private Double endLatitude;
     private Double endLongitude;
-    private Double mileage;
+    private BigDecimal mileage;
 
 }
